Do overcome CORS error
The CORS section was earlier added only to doOption, so even though the `webserver.http.cors.enabled` is set to `true` and `webserver.http.cors.origin` value is provided in config file, the get and post methods used to fail.  Apart from this an header `Allow Control Credentials` was expected to be `true` but the response we were getting was null or `''`
This has resolved my CORS issue I face.